requirements.txt
# Dependências de produção
pandas==2.1.3
xlsxwriter==3.1.9
requests==2.31.0
urllib3==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
workalendar==17.0.0
psycopg2-binary==2.9.9  # PostgreSQL para telemetria (opcional)
# Ferramentas de segurança
bandit==1.7.5              # SAST para Python
pip-audit==2.6.1           # Scanner de vulnerabilidades
safety==3.0.1              # Alternativa a pip-audit

# Ferramentas de qualidade de código
black==23.12.1             # Formatação
isort==5.13.2              # Ordenação de imports
flake8==7.0.0              # Linter
mypy==1.8.0                # Type checker

# Testes
pytest==7.4.3              # Framework de testes
pytest-cov==4.1.0          # Cobertura de testes